+++ projects/frontend/es-app/src/MainRunner.h
@@ -30,6 +30,7 @@ class MainRunner
   , private IFileSystemWatcherNotification
   , public IHardwareNotifications
   , private ILongExecution<HardwareTriggeredSpecialOperations, bool>
+  , private IRomFolderChangeNotification
 {
   public:
     //! Pending Exit
@@ -94,6 +95,13 @@ class MainRunner
 
     //! Window reference
     ApplicationWindow* mApplicationWindow;
+
+    //! Known added devices
+    HashSet<std::string> mAddedDevices;
+    //! Known empty added devices
+    HashSet<std::string> mAddedEmptyDevices;
+    //! Known removed devices
+    HashSet<std::string> mRemovedDevices;
 
     /*!
      * @brief Reset last exit state
@@ -289,4 +297,26 @@ class MainRunner
      * @param executablePath Path to current executable
      */
     static void SetLocale(const std::string& executablePath);
+
+    /*
+     * RomFolderChangeNotification implementaton
+     */
+
+    /*!
+     * @brief Notify a new rompath has been added
+     * @param root rompath added
+     */
+    void RomPathAdded(const DeviceMount& root) override;
+
+    /*!
+     * @brief Notify an existing reompath has been removed
+     * @param root rompath removed
+     */
+    void RomPathRemoved(const DeviceMount& root) override;
+
+    /*!
+     * @brief Notify a device has been added with no valid rom path found
+     * @param deviceRoot Device mount point
+     */
+    void NoRomPathFound(const DeviceMount& deviceRoot) override;
 };